refactor: nearShop 조회 User 패키지 Shop 패키지로 이동
